0.0.0.0.0.2

+++ amazon.py
@@ -9,6 +9,8 @@ import urllib
 import fileinput
 import time
 import re
+import random
+import os
 
 class AmazonBook:
     isEbook = False
@@ -24,11 +26,11 @@ class AmazonBook:
     asin = ''
     publishDate = ''
 
-def getBS_AmazonCOM(pages=10):
-    bs_url = 'https://www.amazon.com/Best-Sellers-Womens-Pumps/zgbs/fashion/679416011'
+def getBS_AmazonCOM(pages=1):
+    bs_url = 'https://www.amazon.com/Best-Sellers-Beauty-Bathing-Accessories/zgbs/beauty/11056491'
 
     for page in range(pages):
-        #time.sleep (3)
+        time.sleep (3)
         r = requests.get(bs_url + '/ref=zg_bs_pg_' + str(page + 1) + '?_encoding=UTF8&pg=' + str(page + 1))
         soup = BeautifulSoup(r.text, 'lxml')
         Books = []
@@ -37,7 +39,17 @@ def getBS_AmazonCOM(pages=10):
             b.isEbook = True
             b.asin = json.loads(item.findAll('div', {'class': 'p13n-asin'})[0]['data-p13n-asin-metadata'])['asin']
             b.rank = item.findAll('span', {'class': 'zg_rankNumber'})[0].text.strip()
-            b.name = item.findAll('img')[0]['alt']
+
+            try:
+                b.pr = item.findAll('span', {'class': 'p13n-sc-price'})[0].text.strip('$')
+                b.pri = int(float(b.pr))
+                b.price = str(b.pri)
+            except IndexError:
+                a = random.randint(1, 10)
+                b.price = str(a)
+                print a
+                print 'oops'
+
 
             try:
                 b.author = item.findAll('div', {'class': 'a-row'})[0].text.strip()
@@ -59,15 +71,21 @@ def getBS_AmazonCOM(pages=10):
             Books.append(copy.copy(b))
 
         res = '\n'.join([book.asin for book in Books]) + '\n\t'
-        print res
-        with io.open('asin.txt', 'a+') as f:
+        res_price = '\n'.join([book.price for book in Books]) + '\n\t'
+        with io.open('asin.txt', 'w') as f:
             f.write(unicode(res))
+        f.close()
+
+        with io.open('price.txt', 'w') as f:
+            f.write(unicode(res_price))
         f.close()
 
 def getBS_AmazonIMG():
 
     with open('asin.txt') as f:
-        for url in f:
+        lines = (line.rstrip() for line in f)
+        lines = (line for line in lines if line)
+        for url in lines:
             time.sleep(3)
             bs_url = 'https://www.amazon.com/dp/'+ url.strip()
             print '\n' + 'New item'
@@ -77,25 +95,38 @@ def getBS_AmazonIMG():
             soup = BeautifulSoup(r.text, 'lxml')
             img = []
 
-            imgs = soup.find("div", {"id": "imgTagWrapperId"}).find("img")
-
+            try:
+                imgs = soup.find("div", {"id": "imgTagWrapperId"}).find("img")
+            except AttributeError:
+                print '\n' + 'Images loading fail. Maybe block from Amazon?' + '\n'
             data = json.loads(imgs["data-a-dynamic-image"])
-            filename = url.strip() + '.jpg'
             x = 0
 
             for img in data:
                 if x == 1:
                     print 'We already got image, skipping ' + url.strip()
+                    exit
+                else:
+                    fl = random.randint(1,18)
+                    add_fl = str(fl)
+                    filename = str(fl) + '.jpg'
+                    print filename
+                    x = x + 1
+                    if os.path.exists(filename):
+                        print 'OwowowWOwowowowowowowowowo'
+                        print 'Saving with new name'
+                        filename_add = add_fl + '(' + add_fl + ')' + '.jpg'
+                        urllib.urlretrieve(img, filename_add)
+                        print 'Saved ' + img + ' in ' + filename_add + ' \n'
+                    else:
+                        urllib.urlretrieve(img, filename)
+
+
                     s = open("asin.txt").read()
                     s = s.replace(url.rstrip(), '\r')
                     f = open("asin.txt", 'w')
                     f.write(s)
                     f.close()
-                    exit
-                else:
-                    print 'Saved ' + img + '\n'
-                    x = x + 1
-                    urllib.urlretrieve(img, filename)
 
-#getBS_AmazonCOM()
+getBS_AmazonCOM()
 getBS_AmazonIMG()
